User commit

+++ src/main/java/com/example/demo/service/StudentService.java
@@ -7,6 +7,6 @@ import com.example.demo.entity.Student;
 public interface StudentService{
     Student insertStudent(Student st);
     List<Student> getAllStudents();
-    Optional<Student> getOneStudent(Long id);
-    void deleteStudent(Long id);
+    Optional<Student> getOneStudent(int id);
+    void deleteStudent(int id);
 }
